refactor:extract all class encapsulate in ConfigurationObject class. ConfigurationObject class is removed.
