+++ js/src/main/scala/inrae/semantic_web/driver/ComunicaFederatedStrategy.scala
@@ -1,6 +1,6 @@
 package inrae.semantic_web.driver
 
-import inrae.semantic_web.configuration.ConfigurationObject.Source
+import inrae.semantic_web.configuration._
 import inrae.semantic_web.driver.ComunicaRequestDriver.SourceComunica
 import inrae.semantic_web.event.{DiscoveryRequestEvent, DiscoveryStateRequestEvent}
 import inrae.semantic_web.sparql.QueryResult
