+++ jvm/src/main/scala/inrae/semantic_web/strategy/Rdf4jFederatedStrategy.scala
@@ -1,7 +1,7 @@
 package inrae.semantic_web.strategy
 
 import inrae.semantic_web.exception._
-import inrae.semantic_web.configuration.ConfigurationObject.Source
+import inrae.semantic_web.configuration._
 import inrae.semantic_web.driver.{Rdf4jRequestDriver, Rdf4jSparqlRequestDriver, RequestDriver, RequestDriverFactory}
 import inrae.semantic_web.event.{DiscoveryRequestEvent, DiscoveryStateRequestEvent, Publisher, Subscriber}
 import inrae.semantic_web.sparql.QueryResult
