+++ shared/src/main/scala/inrae/semantic_web/exception/SWStatementConfigurationException.scala
@@ -1,0 +1,4 @@
+package inrae.semantic_web.exception
+
+final case class SWStatementConfigurationException(private val message: String = "",
+                                                 private val cause: Throwable = None.orNull) extends Exception(message,cause)
